Share observability text redaction
